ai_worker/requirements.txt
fastapi==0.121.0
uvicorn[standard]==0.38.0
onnxruntime==1.21.0
pillow==12.0.0
numpy>=1.24,<3
python-multipart==0.0.20
